requirements.txt
pydantic
openai
flask
python-dotenv
pymongo
langdetect
pinecone